requirements.txt
streamlit
pandas
numpy
scikit-learn
mlflow
pymongo
python-dotenv
plotly
shap
certifi
joblib
